refactor: renamed to fsUtils
